encoding: Encode TA as unsigned and check validty against GSM48_TA_INVALID
According to 3GPP TS 44.018 sec 10.5.2.40, Timing Advance value is 8 bit
and range is 0-63 (0-219 on GSM400). unsigned value (uint8_t) is used
everywhere else, so avoid using a signed one here, and simply check for
GSM48_TA_INVALID here.ch we use everywhere else to initialize when the
value is not known. Ideally we should check for value based on band.

Change-Id: I82b13561d0fe5ebafb5c3a8b9a501045c29809bc